Bump Version 1.2.834

+++ source/ddb/methods/table_drop.py
@@ -6,7 +6,11 @@ def method_drop_table(context, meta):
     context.info("Drop Table")
     try:
         table=get_table(context,meta)
+        if table==None:
+            raise Exception("Table not found")
         
+        print table.data.name
+        print table.data.database
         results = context.database.drop_table(table_name=table.data.name,database_name=table.data.database)
         # TODO Error Handeling
         return query_results(success=results)
